+++ libweva/tests/main.cpp
@@ -419,6 +419,7 @@ void test_abi_placeholder_and_selection_pseudos();
 void test_abi_position_sticky();
 void test_abi_at_import();
 void test_abi_at_import_paths();
+void test_abi_stylesheet_asset_origins();
 void test_abi_scroll_snap();
 void test_abi_cursor();
 void test_abi_range_click_at_edge();
@@ -1005,6 +1006,7 @@ int main() {
     test_abi_position_sticky();
     test_abi_at_import();
     test_abi_at_import_paths();
+    test_abi_stylesheet_asset_origins();
     test_abi_scroll_snap();
     test_abi_cursor();
     test_abi_range_click_at_edge();